renamed js files
