+++ lib/bike.js
