+++ lib/bus.js
