+++ lib/index.js
@@ -1,10 +1,9 @@
-const Frog = require('../lib/frog');
-const Car = require('../lib/cars');
-const Bus = require('../lib/buses');
-const Bike = require('../lib/bikes');
-const Zone = require('../lib/zones');
-const Pad = require('../lib/pads');
-// const Pads = require('../lib/pads');
+var Frog = require('../lib/frog');
+var Car = require('../lib/car');
+var Bus = require('../lib/bus');
+var Bike = require('../lib/bike');
+var Zone = require('../lib/zone');
+var Pad = require('../lib/pad');
 
 var canvas = document.getElementById('world');
 var ctx = canvas.getContext('2d');
@@ -25,19 +24,21 @@ var bikeGirl = new Bike(200, 63);
 var startZone = new Zone();
 var midZone = new Zone(5, 245, 590, 10);
 var endZone = new Zone(5,5);
-var pad = new Pad();
-var padOne = new Pad(70);
-var padTwo = new Pad(150);
-var padThree = new Pad(230);
-var padFour = new Pad(390);
-var padFive = new Pad(470);
-var padSix = new Pad(550);
-var padSeven = new Pad(70, 30);
-var padEight = new Pad(230, 30);
-var padNine = new Pad(390, 30);
-var padTen = new Pad(550, 30);
+var pads = [];
 
-// var pads = new Pads();
+var spotBottomX = 70;
+var spotTopX = 70;
+var spotTopY = 30;
+for (var i = 0; i < 11; i++) {
+  if (i < 7) {
+    pads.push(new Pad(spotBottomX));
+    spotBottomX += 80;
+
+  } else {
+    pads.push(new Pad(spotTopX, spotTopY));
+    spotTopX += 160;
+  }
+}
 
 document.addEventListener('keydown', function(event) {
   if (event.keyCode === 37) {
@@ -59,18 +60,9 @@ requestAnimationFrame(function gameLoop() {
   startZone.draw(ctx);
   midZone.centerLine(ctx);
   endZone.draw(ctx);
-  // pads.draw(ctx);
-  pad.draw(ctx);
-  padOne.draw(ctx);
-  padTwo.draw(ctx);
-  padThree.draw(ctx);
-  padFour.draw(ctx);
-  padFive.draw(ctx);
-  padSix.draw(ctx);
-  padSeven.draw(ctx);
-  padEight.draw(ctx);
-  padNine.draw(ctx);
-  padTen.draw(ctx);
+  pads.forEach(function(pad) {
+    pad.draw(ctx);
+  });
   car.draw(ctx).moveRight();
   carAlso.draw(ctx).moveRight();
   bus.draw(ctx).moveRight();
